fix warnings

+++ vlib/rand/system_rng.c.v
@@ -204,7 +204,7 @@ pub fn (r SysRNG) intn(max int) int {
 		eprintln('max has to be positive.')
 		exit(1)
 	}
-	return int(r.u32n(max))
+	return int(r.u32n(u32(max)))
 }
 
 // r.i64n(max) returns a pseudorandom i64 that lies in [0, max)
@@ -214,7 +214,7 @@ pub fn (r SysRNG) i64n(max i64) i64 {
 		eprintln('max has to be positive.')
 		exit(1)
 	}
-	return i64(r.u64n(max))
+	return i64(r.u64n(u64(max)))
 }
 
 // r.int_in_range(min, max) returns a pseudorandom int that lies in [min, max)
